add conditional printing for testing purposes
all print and println were replaced with a function called printChecker
This function checks a Final variable PRINT to determine if it should
print the JSON or not. 

This function will be used to debug the spacing/bracket issue as shown
"        _links:        {"
This is output form the console as of current.